requirements.txt
asgiref==3.11.1
Django==6.0.2
django-environ==0.12.0
sqlparse==0.5.5
tzdata==2025.3
Pillow==11.1.0
